[ESP-IDF] Move EspHal from example into src/hal/ESP-IDF
The ESP-IDF HAL used to live in examples/NonArduino/ESP-IDF/main/EspHal.h
as a header-only blob, requiring every user of the example to copy it
into their project. The other in-tree HALs (Arduino, RPi, RPiPico) live
under src/hal/<Platform>/. Move EspHal there to match the convention.

Split the implementation into EspHal.h + EspHal.cpp (the free helper
functions getApbFrequency/spiFrequencyToClockDiv become static-linkage
helpers in the .cpp so the header is safe to include in multiple TUs).

Wire the new sources into the IDF component build in the top-level
CMakeLists.txt and add esp_driver_gpio/esp_timer to REQUIRES so the
HAL .cpp can resolve its IDF driver headers. Update the example main.cpp
include path, README, and keywords.txt accordingly.

No behavior changes - the HAL implementation is byte-for-byte identical
to the previous example version.
