+++ CMakeLists.txt
@@ -13,9 +13,14 @@ if(ESP_PLATFORM)
   # required because ESP-IDF runs cmake in script mode
   # and needs idf_component_register()
 
+  # add the ESP-IDF HAL implementation, which was excluded by the filter above
+  file(GLOB RADIOLIB_ESP_IDF_HAL_SOURCES "src/hal/ESP-IDF/*.cpp")
+  list(APPEND RADIOLIB_SOURCES ${RADIOLIB_ESP_IDF_HAL_SOURCES})
+
   idf_component_register(
     SRCS ${RADIOLIB_SOURCES}
-    INCLUDE_DIRS . src 
+    INCLUDE_DIRS . src
+    REQUIRES esp_driver_gpio esp_timer
   )
 
   return()
