+++ examples/NonArduino/ESP-IDF/README.md
@@ -1,11 +1,10 @@
 # RadioLib ESP-IDF example
 
-This example shows how to use RadioLib as ESP-IDF component, without the Arduino framework. To run in ESP-IDF (or on any other platform), RadioLib includes an internal hardware abstraction layer (HAL). This software layer takes care of basic interaction with the hardware, such as performing SPI transaction or GPIO operations. To run on your chosen ESP microcontroller, you will likely have to customize the example HAL for your specific ESP type.
+This example shows how to use RadioLib as ESP-IDF component, without the Arduino framework. To run in ESP-IDF (or on any other platform), RadioLib includes an internal hardware abstraction layer (HAL). This software layer takes care of basic interaction with the hardware, such as performing SPI transaction or GPIO operations. The ESP-IDF HAL ships with RadioLib in `src/hal/ESP-IDF/EspHal.{h,cpp}` and is built automatically when RadioLib is included as an ESP-IDF component.
 
 ## Structure of the example
 
 * `main/CMakeLists.txt` - IDF component CMake file
-* `main/EspHal.h` - RadioLib HAL example implementation
 * `main/idf_component.yml` - declaration of the RadioLib dependency for this example
 * `main/main.cpp` - the example source code
 * `sdkconfig.defaults` - List of preset configuration option values for ESP-IDF. All other options use default values provided by ESP-IDF.
